Re-implemented the Collections functionality
Was a mashup of my experiments with Ember. I have no re-implemented it (still wip) where we have a service that holds 'global state' being the collections themselves, these are exposed as a getter/setter that will, under the hood, go to local storage to get the latest collections. The Collections component, once loaded, will get the latest collections through the service and then setup the tabs and yield the active collection etc. Then when the active tab is refreshed, it will use the service which will update the collections 'state' which ripples down and re-renders where required. This can probably be improved upon eheavily